requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-ajax-datatable==4.2.0
django-bootstrap-modal-forms==2.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mixed-naive-bayes==0.0.1
mlxtend==0.18.0
mysqlclient @ file:///D:/python39/python38/sppd/yang_diperlukan/mysqlclient-1.4.6-cp38-cp38-win32.whl
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
pycparser==2.20
pydataset==0.2.0
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
statsmodels==0.12.2
tablib==3.0.0
terminado==0.9.5
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
